Refactored `static/`
This resolves http://github.com/redpwn/rctf/issues/56
